Added Open World gamemode

+++ core/src/main/java/dev/ultreon/bubbles/bubble/HardenedBubble.java
@@ -19,7 +19,7 @@ public class HardenedBubble extends BubbleType {
     @Override
     public float getDefense(World world, Rng rng) {
         var val = world.getLocalDifficulty() * 4;
-        return rng.getNumber(val / 4f, 3f * val / 4f, world.getTicks(), 1L);
+        return (float) rng.getNumber(val / 4f, 3f * val / 4f, world.getTicks(), 1L);
     }
 
     @Override
